simplified chat store
